refactored chat area out
