+++ src/types/message.ts
@@ -1,0 +1,5 @@
+export interface Message {
+	id: string;
+	role: "user" | "assistant";
+	content: string;
+}
